Reworks Whack a Witch into Whack a Ghost
Following the discussions in #56, it was agreed that using a witch was not
the best option. This changes the witch sprite to a ghost. The lesson plan
has been updated, and also the sample Scratch project. This change applies
to the en-GB version only (this will have to propagated to other languages
by our translators). Furthermore, the sample PDF at Term 1/example lessons
also needs to be regenerated.