Translation Rock Band to Japanese
Let me explain a point which special attention should be paid to for
Japanese language.
Japanese language is written using a combination of two syllabaries
(hiragana and katakana) and Chinese characters.
Japanese children are familiar with hiragana and katakana.
But it is not very easy for most children to read Chinese characters.
To let them know how to read it, hiragana is attached to difficult
Chinese characters in general.
I am sure it is a good way to make the website childlen-friendly.
You would find some code like following in the ".md" file which I
translated.

<ruby>"Chinese character"<rp>（</rp><rt>"hiragana"</rt><rp>）</rp></ruby>

By using this code, hiragana is attached to Chinese character. I hope
this modification does not cause any errors or troubles.

Best Regards,
mushiva2000